Remove unused storev2 code
