+++ backend/store/postgres/entity_config_store.go
@@ -13,6 +13,7 @@ import (
 	corev3 "github.com/sensu/sensu-go/api/core/v3"
 	"github.com/sensu/sensu-go/backend/store"
 	"github.com/sensu/sensu-go/backend/store/patch"
+	storev2 "github.com/sensu/sensu-go/backend/store/v2"
 )
 
 var (
@@ -31,6 +32,20 @@ func NewEntityConfigStore(db *pgxpool.Pool) *EntityConfigStore {
 		db: db,
 	}
 }
+
+type wrapperWithParams interface {
+	storev2.Wrapper
+	SQLParams() []any
+}
+
+type namedWrapperWithParams interface {
+	wrapperWithParams
+	GetName() string
+}
+
+type namespacedResourceNames map[string][]string
+
+type resourceRequestWrapperMap map[storev2.ResourceRequest]storev2.Wrapper
 
 // CreateIfNotExists creates an entity config using the given entity config struct.
 func (s *EntityConfigStore) CreateIfNotExists(ctx context.Context, config *corev3.EntityConfig) error {
